Add run for player in FirstPersonController.cs
Add run by utilizing Vector2 for calculating current speed. This makes it possible for the player to accelerate up to a certain speed. When no input is detected the player will instantly stop.

Setting a higher acceleration value will make the acceleration go faster due to the multiplication with Time.deltaTime.